refactor : change file name and styles
