+++ app/sign-up/life-rhythm.tsx
