--- app/userDetail/_layout.tsx
@@ -1,5 +1,0 @@
-import { Stack } from "expo-router";
-
-export default function UserDetailLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />;
-}
